+++ package.json
@@ -9,18 +9,24 @@
   "author": "TJ Holowaychuk <tj@vision-media.ca>",
   "repository": "git://github.com/component/package.js.git",
   "dependencies": {
-    "mkdirp": "0.3.4",
-    "superagent": "0.15.7",
-    "batch": "0.3.2",
-    "netrc": "~0.1.3",
-    "debug": "*",
-    "rimraf": "~2.1.4",
-    "proxy-agent": "~0.0.2",
-    "superagent-proxy": "~0.2.0"
+    "graceful-fs": "*",
+    "co": "*",
+    "cogent": "*",
+    "archan": "*",
+    "rimraf": "*",
+    "mkdirp": "*",
+    "netrc": "*",
+    "debug": "*"
   },
   "devDependencies": {
+    "regenerator": "*",
+    "gnode": "*",
     "mocha": "*",
     "should": "*"
+  },
+  "scripts": {
+    "test": "make test",
+    "prepublish": "make"
   },
   "engines": {
     "node": ">= 0.8.0"
